Minor refactoring, move into separate package
